build and test mysql-8.0 with tokudb

+++ mysql-8.0/run.mtr.tokudb.bash
@@ -1,11 +1,35 @@
-for suite in tokudb.add_index tokudb.alter_table tokudb tokudb.bugs tokudb.parts tokudb.rpl tokudb.sys_vars ; do
-    rm -rf var $suite.var
-    ./mtr --suite=$suite --skip-test=fast_up.* --force --retry=0 --max-test-fail=0 --parallel=auto --no-warnings --testcase-timeout=60 --big-test >$suite.out 2>&1
-    mv var $suite.var
+engine=tokudb
+
+for suite in $(basename -a suite/tokudb*); do
+    if [ ! -f $suite.out ] ; then
+        rm -rf var $suite.var
+        ./mtr --suite=$suite --skip-test=fast_up.* --force --retry=0 --max-test-fail=0 --parallel=auto --no-warnings --testcase-timeout=60 --big-test >$suite.out 2>&1
+        mv var $suite.var
+    fi
 done
 
 for suite in funcs iuds ; do
-    rm -rf var engines.$suite.var
-    ./mtr --suite=engines/$suite --mysqld=--default-storage-engine=tokudb --mysqld=--default-tmp-storage-engine=tokudb --mysqld=--plugin-load=tokudb=ha_tokudb.so --mysqld=--loose-tokudb-check-jemalloc=0 --parallel=auto --force --retry=0 --max-test-fail=0 --big-test >engines.$suite.out 2>&1
-    mv var engines.$suite.var
+    if [ ! -f $engine.engines.$suite.out ] ; then
+        rm -rf var $engine.engines.$suite.var
+        ./mtr --suite=engines/$suite --mysqld=--default-storage-engine=$engine --mysqld=--default-tmp-storage-engine=$engine --mysqld=--plugin-load=tokudb=ha_tokudb.so --mysqld=--loose-tokudb-check-jemalloc=0 --parallel=auto --force --retry=0 --max-test-fail=0 --big-test >$engine.engines.$suite.out 2>&1
+        mv var $engine.engines.$suite.var
+    fi
+done
+
+engine=innodb
+
+for suite in $(basename -a suite/innodb*); do
+    if [ ! -f $suite.out ] ; then
+        rm -rf var $engine.var
+        ./mtr --suite=$suite --force --retry=0 --max-test-fail=0 --parallel=auto --no-warnings --testcase-timeout=60 --big-test >$suite.out 2>&1
+        mv var $suite.var
+    fi
+done
+
+for suite in funcs iuds ; do
+    if [ ! -f $engine.engines.$suite.out ] ; then
+        rm -rf var $engine.engines.$suite.var
+        ./mtr --suite=engines/$suite --mysqld=--default-storage-engine=$engine --mysqld=--default-tmp-storage-engine=$engine --parallel=auto --force --retry=0 --max-test-fail=0 --big-test >$engine.engines.$suite.out 2>&1
+        mv var $engine.engines.$suite.var
+    fi
 done
